Inline ClassOrganization>>#removeElement: in ClassDescription
Inline ClassOrganization>>#removeElement: in ClassDescription.

Since the name is not really explicit I renamed it #removeFromProtocols: in class description.

I also improved a little some of the methods calling this one.
The removal of #testProtocolNames is done because there is now a better version in ClassDescriptionProtocolsTest and we don't need this duplication of tests