Merge SHTextStyler and SHRBTextStyler
SHTextStyler is never used and only has one subclass. I porpose to merge both of them for now. 

With this change, we can inherit from RBProgramNodeVisitor instead of using TRBProgramNodeVisitor. 

With this, we have no user of TRBProgramNodeVisitor. I think we should deprecate and remove it in the future because it cause us to have to maintain 2 full API of the AST visitor increasing the complexity of the AST maintenance. Especially since the visitor is still been improved (for example to manage better the comments).
What do you think about this next step?