Disable Epicea debug mode in tests
Epicea has a debug mode to catch errors. This is good for the users to get less noise when there is a bug (and Pharo been a big project, we have bugs for sure). But this is not good for the development of Pharo because we cannot know when an error happens.

The best solution for me would be to enable the debug mode only for released versions. But we don't have thi information currently in the image and this would require to update the bootstrap to add the info in SystemVersion.
Here I'm going for the second best thing I could think of: let the errors pass during the tests so that we can at least catch what the tests are unveiling. 

Iâ¯expect this PR to fail because marcus found some log producing errors. But at least we'll be able to detect them better and fix them before integrating this.