Fix logging of protocol in Opal
When we compile a method we save its sources in the source file. This log includes the protocol. 

Currently, it logs the protocol given to Opal and logs before we classify it. But this is not always the final protocol. In some cases, the classification does not use what is given as parameter.

I propose to classify the method before logging it and to ask the protocol directly to the method instead of using the one given as parameter.