Clean fluid class builder tests
Make sure all tests generates their classes in the same package and unify the removal of the generated code.

This also workaround a bug in RPackage category mess that cause to confuse packages and tags. This just hides the problem, but anyway all assertions showing the problem are commented and this category mess should be resolved in the next few months (I hope)