Reduce dependencies of SessionManager in the core of Pharo
Currently in the KernelGroup of the bootstrap we have multiple packages that are only needed by the SessionManager of Pharo. I want to check if we can extract the SessionManager from the core of the bootstrap and make it load via Hermes instead of Espell.

Here is a first step toward that goal. Now all registrations of KernelGroup classes are done in SessionManager class>>#initializeKernelRegistrations to avoid dependencies to SessionManager.

If this pass, then the only dependency we will still have is SmalltalkImage. But that is for another PR since it is not so simple to cut this dependency.

Redo of #13436