Remove some references to `RPackage class>>orginizer`
We have a tons of way to access the default RPackageOrganizer of the image and Iâ¯removed at least 3 ways already. 
In this process of unification I found yet another way to access it, via `RPackage class>> organizer`. 

This is a first step to unify the access to the default RPackageOrganizer for senders of #organizer. 

This is only a part of it because multiple senders are in external projects and there are more in Pharo but I didn't wanted to do a PR that would be a beast.