This PR simplifies what happens when you try to assign to a pseudvariable.
Right now, we get a special error message 'Assigment to reserved variable'. We need to rename the term anyway (they should not be "reserved", and in documentation we use PseudoVariable).

While starting to look at that, I wonderd: does it really add anything to have the special error notice? There is already 'Assignment to read-only variable', and these variables are just one case of those.

This PR thus simplifies to remove the special check, #analyzeWrite:by: now just checks for #isWritable