Deprecate SHRBTextStyler>>#format:
The #format: method does nothing except returning the argument. This is really really really bad for code understanding. I would expect `styler format: aText` to style the text, but no, it does nothing. 
I propose to deprecate this method. And it had senders in the image! I don't know if the senders where expecting this method to do something, but it clearly was not doing anything in the current state of the image.