Bring back coverage test
One coverage tests has been disabled after failing randomly on the CI. The problems comes from some VM optimizations on the VM.
In linux, special selectors  are running in the interpreter so the messages Point >> #x and Point >> #y are never executed.
In OSX, the method is compiled to machine code and the messages are sent.

To have a more robust test, I excluded the special selectors from the list of methods to cover