Simplify Metacello testing
Metacello has some classes present to speed up the testing. But I noticed that if I remove some of them, tests are still been executed in less than a second. 
I think this specific optimization is nos useful because it makes the code more complex for almost no gain. 

I propose to remove it. This has the impact of removing the whole content of Metacello-TestsCommonMC package, so we can remove it alltogether.