Reduce Monticello coupling in RPackage
My goal is to kill all dependencies of RPackage to Monticello. Here is a new step.

Instead of making RPackage listen to the MCWorkingCopyCreated announcement, I make Monticello directly ensure we have the package in the system.

I also provided a small cleaning by deprecating #packageExactlyMatchingExtensionName: in favor of #packageNamedIgnoreCase:ifAbsent:. The reasons are that first I don't find the name really revealing. For me "extension" is about extension methods, but here we get a package name as parameter and this has nothing to do with extension methods. And also there is no need to have multiple methods doing the same things with a slight implementation change. This improves the consistance of RPackage.