Do not poluate Monticello cache in tests
MCWorkingCopyTest is generating content in the tests in the Monticello cache and does multiple cleanings to remove them. 

In order to not impact the filse of the user and also to speed up the tests when we have a big cache, Iâ¯propose to put the cache in memory during the execution of the tests. On my machine tests went from 3min to 12seconds.