This fixes #7218
When defining nil subclasses with non-fluid, at some point the ST80 idea of having class creation methods on nil was replaced by doing a subclass of ProtoObject and then changing the superclass.
This was done as due to the explosion of creation methods, the old way did not scale.

With fluid, we are back with nil subclass. Why not do this for non-fluid, too? We just need to support the old method and the one printed now.