Packages should know their organizer
The RPackage instances do not directy know their organizer, instead they are pointing a class variable that can be updated with a weird method. This leads to really weird behavior. For example if you ask a new organizer instance to register a package, it end up registering it in another PackageOrganizer instance. 
The goal of the PRâ¯is to allow to save the organizer in an instance variable. 

Since we have a lot of ways to register a package, this will not set the variable in all casses but this is a start. (I want to do small steps since it impacts **a lot** of things)