Deprecate RPackageOrganizer>>#ensureExistAndRegisterPackageNamed:
RPackageOrganizer>>#ensureExistAndRegisterPackageNamed: is bad because it does not register the new category in the SystemOrganizer (and even if I want to kill the SystemOrganizer, we still need it until everything using it is migrated) and it does not launch any PackageAdded event, and this is necessary if we want to remove the CategoryAdded event.

Iâ¯propose to deprecate it and to make the users use #registerPackageNamed: instead (this method also does the check that the package exists).

Requires #14110