RPackage: Removing category should remove the tags
Currently we have the SystemOrganizer and the RpackageOrganizer that have duplicated info and are synchronized. But this synchronization have some bugs and one of the is that #removeCategory: does not remove RPackageTag when the category is not a package.
This PR fixes that to be sure we remove the RPackageTag also.

This is necessary to be able to migrate Epicea from category management to Package/Tags management (and kill slowly categories)