Demoting to package with tag should keep the organizer
I started to make sure package know their organizer. Here I am updating the demotion to a package with tag to keep the organizer.

I also started to replave `self packageClass` by `RPackage` because there is no reason to not have a high coupling on RPackage in the packages RPackage and RPackage-Tests and it makes the code more readable and easier to browse.