Reference RPackage in RPackage packages
There is no reason to not directly reference RPAckage class in RPackage packages. There is nothing bad about having a high coupling there.
This will help me with the retro engineering for my future changes to just be able to browse the references to RPackage without to have to look at the senders of #packageClass on top of it.