Refactor: new flow for move-to-class-side-driver supporting bulk refactoring
Now this has better flow when user supplies multiple methods to move to class side.
New flow aggregates all warnings and errors and changes. If there are errors
pop-up is shown. If there are warnings a yes/no pop-up is shown whether user
wants to continue. In the end changes are shown to the user.