Clean #affectsMethodsTaggedWith:
This PR does multiple cleanings on #affectsMethodsTaggedWith:
- Rename it into #affectsMethodsInProtocol:
- Rename variables from tagName to protocol
- Move from the tag API to the Protocol API
- Clean some of the methods

Before this method was excluding Protocol unclassified and extensions but I have the impression that this is never called with those protocol as parameter. Since this was added on System-Announcement and not in Calypso packages, I think we should not do this kind of sorting of what we accept to be impacted or not. We should check the protocol independently of what is the used protocol.