Add test on installing a class in a non default environment
Adding a test about adding a class in a non default environment because I didn't see this done in the image and it made it harder for me to find how to do that.

The problem is that we currently can install a package in another environment but the package organizer it ends up added to is the wrong one! It is always the one of the system and not the one we ask :(
For now I added assertions that I skipped but it is important we are able to make them pass at some point if we want to be able to have tests generating code in the non default environment to not dirty the image.