Simplification of the package validation
Little cleaning while I am trying to understand how RPackage is working before cleaning it further. 
There is a bad hack I found during this session of code reading and I added a flag and comment to explain it. My goal is to remove this hack in the future but it is not yet possible.