Clean package organizer testing
Here is a cleaning of the testing protocol of RPackageOrganizer.

- #hasRegistered has no sender and is removed
- #includesPackage: and #includesPackageNamed: are deprecated
- #hasPackage: has been introduce

The renaming is due to the fact that most other parts of the MM are using #hasX: instead of #includesX: and I want to unify things. 

I did other cleanings for example on refactoring preconditions.
If I have the courage I'll push more things to this PR such as a test for this method or other simplifications in refactorings related to this.