Fix the preconditions related to packages in refactorings
This PR includes multiple cleanings in refacotrings related to packages.

- Use #hasPackage: instead of #includesPackage(Named): that will be deprecated
- Remove useless catch blocks that cannot be triggered
- Remove usage of validatePackageDoesNotExist: because this should be a private method of RPackageOrganizer and use #hasPackage: instead with a better error message
- Do not throw refactoring errors but use #errorString in the preconditions
- Use #packageOrganizer instead of `RPackage organizer` because the later will be deprecated soon 

There are still some things I am unsure about like the usage of #packageOrganizer because it will return the package organizer of the system. Maybe later we should access the packageOrganizer throught the environment? BUT! This should not change the current behavior so this is for a future change. (I'll probably not do it myself, or not before a long time sorry).

This depends on https://github.com/pharo-project/pharo/pull/14232. If you want to review this change before the PR is integrated, you can check the latest commit.