Simplification in Monticello
Monticello contains some really complex code dealings with protocols and packages. In order to simplify the protocol and packages management in the image, it is necessary to do some cleaning in Monticello as well.
In order to better understant it and also to keep my effort of removing the "managers" vocabulary, here I propose a simplification of the code. 

I'll keep pushing this in further PR but with the complexity of the code and how central this is I prefer do small changes.