Clean MCWorkingCopy
This changes finish 3 work that were started on Monticello recently.

The first is to use protocols instead of protocol names were it could simplify the code.

The second was to rename "manager" into "workingCopy" in the code to unify things and make the code more understandable. 

The third was to reduce the "category" dependency of RPackage since we want to remove this in the future. 

The cherry on the cake is that I simplified a lot some parts of the impacted code. Especially #managersForCategory:do: (now called #workingCopiesForExtensionProtocol:do:). And this was already simplified in a recent PR! If you check the code in Pharo 11 and the code now, things are much simpler.