Update bootstrap scripts
This changes update some API in the bootstrap scripts to not use deprecated calls. 

But I see in multiple of those scripts some code that I think cannot work in the current Pharo image. I guess we never use them. I wonder what we should do? Remove them since we do not use them and they are not working anymore? Keep them? But in that case it would be good that the CI runs them to ensure they are up to date..