Introduce RPackageOrganizer>>ensurePackage:
We have a ton of ways to create packages in Pharo and each of them have problems (bad names, inconsistant behaviors, raising or not announcements, ...). I plan to clean that and as a first step I'm introducing a method that should be the main entry point for package creation in the future because I did not find anything in the current API that seems good.

Here is a list of problems:
- #createPackageNamed: it raises an error if the package already exist, but most of the time we just want to "ensure" it exists and we don't want an error if it does.
- #registerPackage: The name "register" is not coherant with the rest of the MM
- #registerPackageNamed: The name "register" is not coherant with the rest of the MM
- #basicRegisterPackage: does not announce anything
- #checkPackageExistsOrRegister: has the category black magic and a bad name 
- #ensureExistAndRegisterPackageNamed: has a bad name and is not coherant with the clean way to register a package 

And some other ways are just dealing with the category black magic.

Once this method is integrated I'll try to deprecate most of the other methods to make this the main entry point for package creation.