Deprecate RPackageOrganizer>>#registerPackageNamed:
Use #ensurePackage: instead because the previous name was using "register" vocabulary that is not consistant with the other parts of the MM and also it was not revealing the fact that it would return an existing package if there was one.

I also removed some unused statements in some tests and I replaced some users of #createPackageNamed: to use #ensurePackage: in places where we do not care if the package already exists.