Remove #colorTransform: from BalloonEngine
BalloonEngine has a variable called #colorTransform: and some code managing it, but this code invoke a non existing method in the case the color is not nil. This means that the code cannot work if the variable is set. 
I propose to remove it. 

WARNING: My image acts weird when I do this change but I would like to see how a bootstraped image act. Maybe this is just due to the migration of the instances in my image.