Deprecate ClassReorganized
ClassReorganized is an announcement that is fired when we add a protocol, remove a protocol or rename a protocol. But this is useless since we have ProtocolAdded, ProtocolRemoved and ProtocolRenamed events!
I propose to remove it because it is really complex to make sur we do not fire it mulitple times, for example when we classify a method that was the last one in its protocol inside a new protocol.

This should fix https://github.com/pharo-project/pharo/pull/14264