Merge addition of class definition in rpackage
One of my goal is to make sure packages use real classes instead of class names. Here is a small step toward that goal: I merged #addClassDefinition: and #addClassDefinitionNamed:

Next step is to merge removeClassDefinition: and #removeClassDefinitionNamed: but this is not as easy since this is used by the system.