Improve CompiledMethod>>#isExtension
In my RPackage cleanings I want to remove all the dictionaries of names to manipulated the MM directly to get the right informations. This would simplify a lot the code and would remove all the bugs linked to synchronization. In order to do that without losing too much performance I want to speed up some operations first.

In this change I speed up CompiledMethod>>#isExtension and also I fixed a bug in it where every method that was removed from a class was answering true to #isExtension. Now it answers false since the method is unpackaged.
I also added a tests to ensure this behavior.

Here are some bench:

```st
[ SystemNavigation default methods do: [ :m | m isExtension ] ] bench

"Before the change: 30 iterations in 5 seconds 13 milliseconds. 5.984 per second"

"After the change: 10 iterations in 5 seconds 436 milliseconds. 1.840 per second"
```