Class are created with the wrong environment
A class knows the environment it is installed in (Class>>#environment). But in fact this is not working because the environment set to the class is the building environment and not the installing environment.

This change adds an assertion in a test to show this bug and propose a fix by moving the instralling environment from the class installer to the class builder and using this environment as the environment of the class.