Clean method compilation of Monticello
Monticello reimplement itself the method installation trying to use the compiler only to compile the method and then managing itself the installation in the class and the change file update.

This seems like some useless duplication. Here is a try to use Opal compiler directly for the method installation to be sure everything is coherent and that there is no bug in what Monticello does. 

This should fix some vanishing protocol creations or other missing announcements.