Forbid to have an extension method in the package owning the class
It is currently possible to have a method that is an extension method that is in the same package as its class. This creates weird behaviors were some tools consider it an extension method and some other consider it not an extension method.

To make thing more coherent I propose that we forbid to have this by classifying such methods under the unclassified protocol instead.