Simplify class removal in packages
RPackage contains a lot of methods to update its caches. A bunch of them are there for the method #definedMethodsShouldBecomeExtensionWhenRemovingClass: to update the cache of the methods.

But this method is only called by two methods:
- One that remove all methods right after (making the update of the caches useless)
- One that is called only by tests

Since all those updates are not necessary and one of my mid-term goal is to remove the selectors caches, I'm proposing to just remove all this mess to make the system a little bit simpler.