Clean method removal of RPackage
This is a first step at the cleaning of the method removal of RPackage. 

This is still quite bad but it is a first step a cleaning this. The change make the method removal work as the method addition and reduces the usage of the cache variables of RPackage a little. My mid term goal is to kill some of those caches.