Simplify class addition assuming classes are created empty
Some code in RPackage assumes that methods can be added to a package before the classes. But we never get in this case since classes are created. empty.
I propose to simplify this code by removing tests showing a behavior never happening and by removing this complexe update code.

Three things to note:
- This was one of the reason preventing some changes to improve the tests. 
-  I plan to do another iteration after this one on class addition to make it simpler once again.
- This reduces the usage of the selectors caches I want to remove in RPackage.