Fix bug of method returning a text as source code
In the past, we were logging a method before installing it in the system. This was not good because the installation could change some values of the method such as the protocol and the logged info was wrong.

Now we are installing the method and then we log it. The problem that arise is that the installation of the method might need the source code. In that case, instead of reading the source code from the change file, we use a property saved in memory. But this property was wrong in case we provided a text to the compiler because it stored the text instead of the string.

I propose to convert the text as a string in the setter of the compiler.

Fixes #14355