Fix bug with reset protocol of Calypso
Calypso provides a protocol editor at the bottom of the editor pane but using the reset button does not work correctly. 

This was due to a hack in the classification of methods preventing to unclassify a method. Thus "nil" was used to unclassify a method but this did not work in all cases. I recently removed this hack and we can now fix more easily this bug. 

I also renamed ClyProtocolAndPackageEditorMorph into ClyProtocolEditorMorph.