Do not add classes through announcements
A new step toward getting away from the synchronization of RPackageOrganizer through announcements. If this change works well, the new classes will not be added via announcements.

This would also be a step toward fixing some failing tests in Spec that are due to the face that some part of the system are trying to generate silently some code suspending announcements, but since RPackage is updated via announcements, it leaves it in an unsynchronized state.