RPackage: Clean bootstrap initialization
There is an initialization of packages in the bootstrap that is redoing some code in it own way instead of using the RPackage classic API. This make it harder to clean the API. 

In the past weeks I tried to unify some of the behavior. Now let's see if the class import can be unified.