RPackage: Use real classes for extension caches
In RPackageOrganizer we have a cache of the classes and the packages extending those classes to speed up the search of packages of the extension methods.

This cache was using a class name and a set of packages. This change updates it to store and identity set of the packages under the real classes and not the names. 

The update code happening on class renaming is starting to be much smaller and easier with all the recent changes