Save real classes in RPackage
RPackage is currently keeping a list of class names to know the classes it contains. This means that we often need to lookup in the system dictionary to know the real classes.

I propose to save the real classes instead since it is easier to ask to a class its name than to lookup the class in the system.

I also make the collection containing the classes a IdentitySet since instance of classes should stay the same and it is faster.

Lastly I removed a test that was really not useful since it tests a behavior that should never happen in the system (and if someone does that, then it's his problem, not RPackage problem)

Redo of #14369