RPackage: Clean package tag API
This change cleans a little the Package tag API and mostly remove methods with no sender and that do not make sense.
For example mulitple methods are about extension methods or protocols while those are managed at the level of the package and not at the level of the tag.