Calypso: Prevent the recategorization of removed methods
Calypso is listening to method changes to update the method pane. But doing this it forces the recategorization of methods in case the method has no protocol. This is happening when we explicitly remove a method from a protocol during a method removal. 

This change make sure we do not update the protocol of a method if its protocol is nil