RPackage: #packageTag should always return a tag
ClassDescription>>#packageTag has a guard to return nil if the class is in the default tag. This can cause some problems and make the RPAckage code more complex. 

I propose to always return a tag and is #isRoot instead of #ifNil: to manage the root tag. This will simplify future changes