Monticello: First step at unifying metohd installation with the kernel
Monticello redid the whole method installation code that we have in the image and this for two reasons:
- When loading a package, we want to have all the classes and methods of the package loaded before executing the #initialize method
- The method added announcement has a property added by Monticello

But this duplication come at a cost. Over the years some changes were made in the Kerned, when adding the Trait support for example, and that was not followed in the Moticello MethodAddition. I would like to remove this duplication to avoid such cases.

Here is a first step introducing a new way to group the method announcements. Instead of executing the method announcing the changes after all method installation, I am suspending all announcements during the installation and I send the announcements as a batch later.